requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib
python-dotenv
torch
numpy
pandas
requests